requirements.txt
matplotlib>=3.8.0
numpy>=1.26.0
Pillow>=10.0.0
PyYAML>=6.0.0
torch>=2.2.0
tqdm>=4.66.0
